[hydro] Remove VolumeElementIndex type (#15933) There are four index types associated with the two mesh types. This removes the fourth of the four indices.
In this case "removal" means:
  1. Removing all usages of VolumeElementIndex in Drake C++ code.
  2. Removing the python bindings.
  3. Keep the type, but change it to be an alias to an int. This will allow
     us to defer updating Anzu until *all* index types have been removed.
     At that point, we'll remove the int aliases and update Anzu.
     - TypeSafeIndex allows for index vs unsigned int comparisions. This
       alias will break this functionality. Where necessary, those signed-
       unsigned comparisons are also updated.

+++ geometry/proximity/mesh_field_linear.h
@@ -217,8 +217,8 @@ class MeshFieldLinear {
     if (gradients_.size() == 0) {
       return Evaluate(e, this->mesh().CalcBarycentric(p_MQ, e));
     } else {
-      DRAKE_ASSERT(e < gradients_.size());
-      DRAKE_ASSERT(e < values_at_Mo_.size());
+      DRAKE_ASSERT(e < static_cast<int>(gradients_.size()));
+      DRAKE_ASSERT(e < static_cast<int>(values_at_Mo_.size()));
       return gradients_[e].dot(p_MQ) + values_at_Mo_[e];
     }
   }
@@ -321,7 +321,7 @@ class MeshFieldLinear {
   }
 
   T CalcValueAtMeshOrigin(typename MeshType::ElementIndex e) const {
-    DRAKE_DEMAND(e < gradients_.size());
+    DRAKE_DEMAND(e < static_cast<int>(gradients_.size()));
     const typename MeshType::VertexIndex v0 = this->mesh().element(e).vertex(0);
     const Vector3<T>& p_MV0 = this->mesh().vertex(v0);
     // f(V₀) = ∇fᵉ⋅p_MV₀ + fᵉ(Mo)
